100% Done

+++ Controllers/DjMusicasInPlayListsController.cs
@@ -7,6 +7,7 @@ using System.Linq;
 
 namespace PayAndPlay.Controllers
 {
+    // CRUD Criado para o DJ gerenciar as musicas em suas PlayLists, onde ele pode adicionar, editar e remover musicas de suas PlayLists, MVC Controller
     public class DjMusicasInPlayListsController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -19,36 +20,63 @@ namespace PayAndPlay.Controllers
         // GET: MusicasInPlayLists
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.TmusicaInPlayLists.Include(m => m.Musica).Include(m => m.PlayList).Where(p => p.PlayList.DJId == int.Parse(HttpContext.Session.GetString("ID")));
-            return View(await applicationDbContext.ToListAsync());
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
+            {
+                var applicationDbContext = _context.TmusicaInPlayLists.Include(m => m.Musica).Include(m => m.PlayList).Where(p => p.PlayList.DJId == int.Parse(HttpContext.Session.GetString("ID")));
+                return View(await applicationDbContext.ToListAsync());
+            }
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // GET: MusicasInPlayLists/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    TempData["Message"] = "Error: Musicas na PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            var musicaInPlayList = await _context.TmusicaInPlayLists
-                .Include(m => m.Musica)
-                .Include(m => m.PlayList)
-                .FirstOrDefaultAsync(m => m.ID == id);
-            if (musicaInPlayList == null)
+                var musicaInPlayList = await _context.TmusicaInPlayLists
+                    .Include(m => m.Musica)
+                    .Include(m => m.PlayList)
+                    .FirstOrDefaultAsync(m => m.ID == id);
+                if (musicaInPlayList == null)
+                {
+                    TempData["Message"] = "Error: Musicas na PlayList não encontrada!";
+                    return NotFound();
+                }
+
+                TempData["Message"] = "Success: Musicas na PlayList encontrados!";
+                return View(musicaInPlayList);
+            }
+            else
             {
-                return NotFound();
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
             }
-
-            return View(musicaInPlayList);
         }
 
         // GET: MusicasInPlayLists/Create
         public IActionResult Create()
         {
-            ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome");
-            ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome");
-            return View();
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
+            {
+                ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome");
+                ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome");
+                return View();
+            }
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // POST: MusicasInPlayLists/Create
@@ -58,33 +86,52 @@ namespace PayAndPlay.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ID,MusicaId,PlayListId")] MusicaInPlayList musicaInPlayList)
         {
-            if (ModelState.IsValid)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                _context.Add(musicaInPlayList);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(musicaInPlayList);
+                    await _context.SaveChangesAsync();
+                    TempData["Message"] = "Success: Musica adicionada na PlayList com sucesso!";
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome", musicaInPlayList.MusicaId);
+                ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome", musicaInPlayList.PlayListId);
+                return View(musicaInPlayList);
             }
-            ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome", musicaInPlayList.MusicaId);
-            ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome", musicaInPlayList.PlayListId);
-            return View(musicaInPlayList);
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // GET: MusicasInPlayLists/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    TempData["Message"] = "Error: Musica na PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            var musicaInPlayList = await _context.TmusicaInPlayLists.FindAsync(id);
-            if (musicaInPlayList == null)
-            {
-                return NotFound();
+                var musicaInPlayList = await _context.TmusicaInPlayLists.FindAsync(id);
+                if (musicaInPlayList == null)
+                {
+                    TempData["Message"] = "Error: Musica na PlayList não encontrada!";
+                    return NotFound();
+                }
+                ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome");
+                ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome");
+                return View(musicaInPlayList);
             }
-            ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome");
-            ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome");
-            return View(musicaInPlayList);
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // POST: MusicasInPlayLists/Edit/5
@@ -94,54 +141,75 @@ namespace PayAndPlay.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("ID,MusicaId,PlayListId")] MusicaInPlayList musicaInPlayList)
         {
-            if (id != musicaInPlayList.ID)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id != musicaInPlayList.ID)
+                {
+                    TempData["Message"] = "Error: Musica na PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(musicaInPlayList);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!MusicaInPlayListExists(musicaInPlayList.ID))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(musicaInPlayList);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!MusicaInPlayListExists(musicaInPlayList.ID))
+                        {
+                            TempData["Message"] = "Error: Musica na PlayList não encontrada!";
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    TempData["Message"] = "Success: Musica na PlayList atualizada com sucesso!";
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome");
+                ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome");
+                return View(musicaInPlayList);
             }
-            ViewData["MusicaId"] = new SelectList(_context.Tmusicas, "ID", "Nome");
-            ViewData["PlayListId"] = new SelectList(_context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))), "ID", "Nome");
-            return View(musicaInPlayList);
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // GET: MusicasInPlayLists/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    TempData["Message"] = "Error: Musica na PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            var musicaInPlayList = await _context.TmusicaInPlayLists
-                .Include(m => m.Musica)
-                .Include(m => m.PlayList)
-                .FirstOrDefaultAsync(m => m.ID == id);
-            if (musicaInPlayList == null)
+                var musicaInPlayList = await _context.TmusicaInPlayLists
+                    .Include(m => m.Musica)
+                    .Include(m => m.PlayList)
+                    .FirstOrDefaultAsync(m => m.ID == id);
+                if (musicaInPlayList == null)
+                {
+                    TempData["Message"] = "Error: Musica na PlayList não encontrada!";
+                    return NotFound();
+                }
+
+                return View(musicaInPlayList);
+            }
+            else
             {
-                return NotFound();
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
             }
-
-            return View(musicaInPlayList);
         }
 
         // POST: MusicasInPlayLists/Delete/5
@@ -149,14 +217,23 @@ namespace PayAndPlay.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var musicaInPlayList = await _context.TmusicaInPlayLists.FindAsync(id);
-            if (musicaInPlayList != null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                _context.TmusicaInPlayLists.Remove(musicaInPlayList);
-            }
+                var musicaInPlayList = await _context.TmusicaInPlayLists.FindAsync(id);
+                if (musicaInPlayList != null)
+                {
+                    _context.TmusicaInPlayLists.Remove(musicaInPlayList);
+                }
 
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+                await _context.SaveChangesAsync();
+                TempData["Message"] = "Error: Não tem permissões para aceder a esta pagina!";
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         private bool MusicaInPlayListExists(int id)
