+++ Controllers/DjPlayListController.cs
@@ -6,6 +6,7 @@ using PayAndPlay.Models;
 
 namespace PayAndPlay.Controllers
 {
+    // CRUD Criado para o DJ gerenciar suas PlayLists, onde ele pode adicionar, editar e remover PlayLists, MVC Controller
     public class DjPlayListController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -18,34 +19,60 @@ namespace PayAndPlay.Controllers
         // GET: PlayLists
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))).Include(p => p.DJ);
-            return View(await applicationDbContext.ToListAsync());
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
+            {
+                var applicationDbContext = _context.TplayLists.Where(p => p.DJId == int.Parse(HttpContext.Session.GetString("ID"))).Include(p => p.DJ);
+                return View(await applicationDbContext.ToListAsync());
+            }
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // GET: PlayLists/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    TempData["Message"] = "Error: PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            var playList = await _context.TplayLists
-                .Include(p => p.DJ)
-                .FirstOrDefaultAsync(m => m.ID == id);
-            if (playList == null)
+                var playList = await _context.TplayLists
+                    .Include(p => p.DJ)
+                    .FirstOrDefaultAsync(m => m.ID == id);
+                if (playList == null)
+                {
+                    TempData["Message"] = "Error: PlayList não encontrada!";
+                    return NotFound();
+                }
+
+                TempData["Message"] = "Success: PlayList encontrada!";
+                return View(playList);
+            }
+            else
             {
-                return NotFound();
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
             }
-
-            return View(playList);
         }
-
         // GET: PlayLists/Create
         public IActionResult Create()
         {
-            ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName");
-            return View();
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
+            {
+                ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName");
+                return View();
+            }
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // POST: PlayLists/Create
@@ -55,31 +82,50 @@ namespace PayAndPlay.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ID,Nome,DJId")] PlayList playList)
         {
-            if (ModelState.IsValid)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                _context.Add(playList);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(playList);
+                    await _context.SaveChangesAsync();
+                    TempData["Message"] = "Success: PlayList criada com sucesso!";
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName", playList.DJId);
+                return View(playList);
             }
-            ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName", playList.DJId);
-            return View(playList);
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // GET: PlayLists/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    TempData["Message"] = "Error: PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            var playList = await _context.TplayLists.FindAsync(id);
-            if (playList == null)
-            {
-                return NotFound();
+                var playList = await _context.TplayLists.FindAsync(id);
+                if (playList == null)
+                {
+                    TempData["Message"] = "Error: PlayList não encontrada!";
+                    return NotFound();
+                }
+                ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName", playList.DJId);
+                return View(playList);
             }
-            ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName", playList.DJId);
-            return View(playList);
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // POST: PlayLists/Edit/5
@@ -89,52 +135,73 @@ namespace PayAndPlay.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("ID,Nome,DJId")] PlayList playList)
         {
-            if (id != playList.ID)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id != playList.ID)
+                {
+                    TempData["Message"] = "Error: PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(playList);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!PlayListExists(playList.ID))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(playList);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!PlayListExists(playList.ID))
+                        {
+                            TempData["Message"] = "Error: PlayList não encontrada!";
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    TempData["Message"] = "Success: PlayList atualizada com sucesso!";
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName", playList.DJId);
+                return View(playList);
             }
-            ViewData["DJId"] = new SelectList(_context.Tdjs, "ID", "UserName", playList.DJId);
-            return View(playList);
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         // GET: PlayLists/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    TempData["Message"] = "Error: PlayList não encontrada!";
+                    return NotFound();
+                }
 
-            var playList = await _context.TplayLists
-                .Include(p => p.DJ)
-                .FirstOrDefaultAsync(m => m.ID == id);
-            if (playList == null)
+                var playList = await _context.TplayLists
+                    .Include(p => p.DJ)
+                    .FirstOrDefaultAsync(m => m.ID == id);
+                if (playList == null)
+                {
+                    TempData["Message"] = "Error: PlayList não encontrada!";
+                    return NotFound();
+                }
+
+                return View(playList);
+            }
+            else
             {
-                return NotFound();
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
             }
-
-            return View(playList);
         }
 
         // POST: PlayLists/Delete/5
@@ -142,14 +209,23 @@ namespace PayAndPlay.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var playList = await _context.TplayLists.FindAsync(id);
-            if (playList != null)
+            if (HttpContext.Session.GetString("UTILIZADOR") != "" && HttpContext.Session.GetString("UTILIZADOR") != null && HttpContext.Session.GetString("PERFIL") == "2" && HttpContext.Session.GetString("ADMIN") == "false")
             {
-                _context.TplayLists.Remove(playList);
-            }
+                var playList = await _context.TplayLists.FindAsync(id);
+                if (playList != null)
+                {
+                    _context.TplayLists.Remove(playList);
+                }
 
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+                await _context.SaveChangesAsync();
+                TempData["Message"] = "Success: PlayList removida com sucesso!";
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                TempData["Message"] = "Error: Nao tem permissoes para aceder a esta pagina!";
+                return RedirectToAction("Index", "Login");
+            }
         }
 
         private bool PlayListExists(int id)
